# gateway/docs/swagger.yaml
basePath: /api
definitions:
  handlers.createUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - password
    - role
    type: object
...
  models.SubmitRunRequest:
    properties:
      agent_type:
        type: string
      model:
        type: string
      task:
        type: string
    required:
    - agent_type
    - model
    - task
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Runtime security platform for autonomous AI agents.
  title: AgentShield API
  version: "1.0"
paths:
...
  /redteam/results:
    get:
      description: Return all completed red-team run summaries.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List red-team runs
      tags:
      - redteam
  /redteam/results/{id}:
    get:
      description: Return detailed per-case results for a single red-team run.
      parameters:
      - description: Red-team run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get red-team run by ID
      tags:
      - redteam
  /redteam/run:
    post:
      description: Run all adversarial test cases against every security interceptor.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Trigger red-team run
      tags:
      - redteam
  /runs:
    get:
      description: Return the 100 most recent agent runs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List agent runs
      tags:
      - runs
    post:
      consumes:
      - application/json
      description: Execute an agent task and persist the result.
      parameters:
      - description: Run request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SubmitRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a new agent run
      tags:
      - runs
  /runs/{id}:
    get:
      description: Return a single agent run by its UUID.
      parameters:
      - description: Run UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get agent run by ID
      tags:
      - runs
...
  /users:
    get:
      description: Return all user accounts. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account (admin or viewer). Admin only.
      parameters:
      - description: New user details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID. Cannot delete your own account. Admin only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
  /ws/events:
    get:
      description: Upgrade to a WebSocket connection and receive live security events.
        Pass JWT via ?token= query parameter.
      parameters:
      - description: JWT access token
        in: query
        name: token
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: WebSocket event stream
      tags:
      - events
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"